Add removeEmptyRows and reorganize tableutilites
Because:
-- UPDOG tsvs contain numerous completely empty rows
-- Reorginized TableCleaner, TableSetCLeaner, and TableUtilities to adhere to single responsiblities

* RenamedTableSetUtilities to TableSetCleaner
* Create new class TableUtilities
* Separate TableSetCleaner from TableCleaner methods
* Separate TableSetCleaner and TableCleaner methods from TableUtilities

* Add behavior to remove completely empty lines in tables
* Add test for this behavior

* Simplified BrokenRelationErrorCreator stream

ticket:#579
